T76/T56 Logic IC Test Support

+++ MiniproUI/Minipro/ResponseProcessors/ProgrammerInfoProcessor.swift
@@ -17,10 +17,10 @@ struct ProgrammerInfo {
     let supplyVoltage: String
     let warnings: [String]
 
-    func getFirmwareVersionNumber() -> Int? {
+    func getFirmwareVersionNumber() -> UInt16? {
         let hexRegex = /0x[0-9a-fA-F]+/
         if let match = try? hexRegex.firstMatch(in: firmwareVersion) {
-            return Int(match.0.dropFirst(2), radix: 16)
+            return UInt16(match.0.dropFirst(2), radix: 16)
         }
         return nil
     }
